* solve the #1 issue https://github.com/kumquatt/mqttpacket/issues/1

+++ src/main/scala-2.11/plantae/citrus/mqtt/packet/Header.scala
@@ -2,7 +2,7 @@ package plantae.citrus.mqtt.packet
 
 import scodec.Codec
 
-case class FixedHeader(dup: Boolean = false, qos: Int = 0, retain: Boolean = false)
+case class FixedHeader(dup: Boolean = false, qos: Short = 0, retain: Boolean = false)
 
 object FixedHeader {
   implicit val codec: Codec[FixedHeader] = fixedHeaderCodec
@@ -12,7 +12,7 @@ case class ConnectVariableHeader(
                                   userNameFlag: Boolean,
                                   passwordFlag: Boolean,
                                   willRetain: Boolean,
-                                  willQoS: Int,
+                                  willQoS: Short,
                                   willFlag: Boolean,
                                   cleanSession: Boolean,
                                   keepAliveTime: Int
